// src/lib/api/gen/types/GetStoreTeam.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type GetStoreTeamPathParams = {
    /**
     * @type string
    */
    slug: string;
};

export const membersRoleEnum = {
    owner: "owner",
    admin: "admin",
    staff: "staff"
} as const;

export type MembersRoleEnumKey = (typeof membersRoleEnum)[keyof typeof membersRoleEnum];

export const invitesRoleEnum = {
    admin: "admin",
    staff: "staff"
} as const;

export type InvitesRoleEnumKey = (typeof invitesRoleEnum)[keyof typeof invitesRoleEnum];

/**
 * @description Default Response
*/
export type GetStoreTeam200 = {
    /**
     * @type array
    */
    members: {
        /**
         * @type string
        */
        userId: string;
        /**
         * @type string
        */
        name: string;
        /**
         * @type string
        */
        email: string | null;
        /**
         * @type string
        */
        role: MembersRoleEnumKey;
    }[];
    /**
     * @type array
    */
    invites: {
        /**
         * @type string
        */
        id: string;
        /**
         * @type string
        */
        email: string;
        /**
         * @type string
        */
        role: InvitesRoleEnumKey;
        /**
         * @type string
        */
        expiresAt: string;
        /**
         * @type string
        */
        createdAt: string;
    }[];
};

export type GetStoreTeamQueryResponse = GetStoreTeam200;

export type GetStoreTeamQuery = {
    Response: GetStoreTeam200;
    PathParams: GetStoreTeamPathParams;
    Errors: any;
};